Integrate undeclared required regressions into existing tests
